% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{build_expr_long}
\alias{build_expr_long}
\title{Building long-format expression data}
\usage{
build_expr_long(sce, genes)
}
\arguments{
\item{sce}{A SingleCellExperiment object}

\item{genes}{Character vector of gene names}
}
\value{
A data.frame with cell-level expression values
}
\description{
Building long-format expression data
}
