% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{compute_expr_stats}
\alias{compute_expr_stats}
\title{Computing per-group expression statistics}
\usage{
compute_expr_stats(sce, genes)
}
\arguments{
\item{sce}{SingleCellExperiment object}

\item{genes}{character vector of genes}
}
\value{
data.frame of expression stats
}
\description{
Computing per-group expression statistics
}
