% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{compute_gene_specificity}
\alias{compute_gene_specificity}
\title{Computing gene specificity (Gini score)}
\usage{
compute_gene_specificity(sce, genes)
}
\arguments{
\item{sce}{SingleCellExperiment object}

\item{genes}{character vector of genes}
}
\value{
data.frame of gene specificity
}
\description{
Computing gene specificity (Gini score)
}
