% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{example_se}
\alias{example_se}
\title{Example SingleCellExperiment dataset from Muraro Subset}
\format{
A SingleCellExperiment object with:
\describe{
\item{assays}{logcounts and counts matrices}
\item{colData}{cell metadata including cell type labels}
\item{rowData}{gene metadata}
}
}
\source{
\link{scRNAseq::MuraroPancreasData}
}
\usage{
example_se
}
\description{
A small subset of the Muraro human pancreas dataset that
contains a reduced number of genes and cells for testing.
}
\examples{
data(example_se)
example_se
}
\keyword{datasets}
