% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_scatter}
\alias{plot_scatter}
\title{Scatter plot between genes}
\usage{
plot_scatter(sce, gene_a, gene_b)
}
\arguments{
\item{sce}{SingleCellExperiment object}

\item{gene_a}{First gene}

\item{gene_b}{Second gene}
}
\value{
ggplot object
}
\description{
Scatter plot between genes
}
