% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_violin}
\alias{plot_violin}
\title{Violin plot of gene expression}
\usage{
plot_violin(expr_long)
}
\arguments{
\item{expr_long}{data.frame from build_expr_long()}
}
\value{
ggplot object
}
\description{
Violin plot of gene expression
}
