% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{save_expr_stats}
\alias{save_expr_stats}
\title{Save expression statistics to a file}
\usage{
save_expr_stats(stats_result, output_file)
}
\arguments{
\item{stats_result}{Output from \code{compute_expr_stats()}}

\item{output_file}{File path to save results}
}
\value{
Invisible NULL; file is written to disk
}
\description{
Save expression statistics to a file
}
\examples{
\dontrun{
data(example_se)
genes <- rownames(example_se)[1:5]
res <- compute_expr_stats(example_se, genes)
save_expr_stats(res, "stats.tsv")
}
}
